lumen_web/requirements.txt
Flask==3.0.3
psycopg
python-dotenv
gunicorn
